Refactor: UI data :ui:data 모듈로 이동
