+++ ui/data/src/main/java/com/kappzzang/jeongsan/data/ExpenseItemInput.kt
@@ -1,3 +1,3 @@
-package com.kappzzang.jeongsan.addexpense
+package com.kappzzang.jeongsan.data
 
 data class ExpenseItemInput(var itemName: String?, var itemPrice: Int?, var itemQuantity: Int?)
