+++ ui/data/src/main/java/com/kappzzang/jeongsan/data/ExpenseListViewUIData.kt
@@ -1,4 +1,4 @@
-package com.kappzzang.jeongsan.expenselist
+package com.kappzzang.jeongsan.data
 
 import com.kappzzang.jeongsan.model.ExpenseItem
