+++ ui/expenselist/src/main/java/com/kappzzang/jeongsan/expenselist/ExpenseListBindingAdapter.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.StateFlow
 object ExpenseListBindingAdapter {
     @BindingAdapter("expenseItems")
     @JvmStatic
-    fun attachExpenseList(recyclerView: RecyclerView, items: StateFlow<ExpenseListViewUIData>?) {
+    fun attachExpenseList(recyclerView: RecyclerView, items: StateFlow<com.kappzzang.jeongsan.data.ExpenseListViewUIData>?) {
         items?.let {
             (recyclerView.adapter as? ExpenseListAdapter)
                 ?.submitList(
