+++ ui/expenselist/src/main/java/com/kappzzang/jeongsan/expenselist/ExpenseListViewModel.kt
@@ -2,6 +2,7 @@ package com.kappzzang.jeongsan.expenselist
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.kappzzang.jeongsan.data.ExpenseListViewUIData
 import com.kappzzang.jeongsan.model.ExpenseListResponse
 import com.kappzzang.jeongsan.model.ExpenseState
 import com.kappzzang.jeongsan.usecase.GetCurrentGroupInfoUseCase
